fixed signed/unsigned issues with GrowingNetwork2D::edgeArcDistance

+++ growingnetwork2d/growingnetwork2d.cpp
@@ -93,7 +93,7 @@ void GrowingNetwork2D::grow(long int n){
 long int GrowingNetwork2D::edgeArcDistance(Vertex* a, Vertex* b){
 	
 	// find the minimum of distance in ascending order vs. descending order of index
-	return min(abs(indexOf(b) - indexOf(a)), N - abs(indexOf(b) - indexOf(a)));
+	return min((unsigned long int)abs(indexOf(b) - indexOf(a)), (unsigned long int)N - abs(indexOf(b) - indexOf(a)));
 	
 }
